pyproject.toml
[project]
name = "pfe-dematis-implem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "audioread>=3.1.0",
    "boto3>=1.42.1",
    "ipywidgets>=8.1.8",
    "jupyter>=1.1.1",
    "jupyterlab>=4.5.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.7",
    "mpi4py>=4.1.1",
    "notebook>=7.5.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "scikit-maad>=1.5.1",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "soundfile>=0.13.1",
    "streamlit>=1.51.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "umap-learn>=0.5.9.post2",
]
